# Microserviciu_2/OpenAPIDocumentation_M2.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Academia Moodle M2 cu MongoDB",
        "description": "API pentru Academia Moodle al doilea microserviciu cu MongoDB",
        "version": "1.0.0"
    },
    "paths": {
        "/evaluare/": {
            "post": {
                "summary": "Create Evaluare",
                "operationId": "create_evaluare_evaluare__post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Evaluare"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/materiale/": {
            "post": {
                "summary": "Create Material",
                "operationId": "create_material_materiale__post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Materiale"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/evaluare/{disciplina}": {
            "get": {
                "summary": "Get Evaluare",
                "operationId": "get_evaluare_evaluare__disciplina__get",
                "parameters": [
                    {
                        "name": "disciplina",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Disciplina"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "summary": "Update Evaluare",
                "operationId": "update_evaluare_evaluare__disciplina__put",
                "parameters": [
                    {
                        "name": "disciplina",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Disciplina"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Evaluare"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/materiale/{disciplina}": {
            "get": {
                "summary": "Get Materiale",
                "operationId": "get_materiale_materiale__disciplina__get",
                "parameters": [
                    {
                        "name": "disciplina",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Disciplina"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "summary": "Update Materiale",
                "operationId": "update_materiale_materiale__disciplina__put",
                "parameters": [
                    {
                        "name": "disciplina",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Disciplina"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Materiale"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Evaluare": {
                "properties": {
                    "disciplina": {
                        "type": "string",
                        "title": "Disciplina"
                    },
                    "probe": {
                        "items": {
                            "$ref": "#/components/schemas/Proba"
                        },
                        "type": "array",
                        "title": "Probe"
                    }
                },
                "type": "object",
                "required": [
                    "disciplina",
                    "probe"
                ],
                "title": "Evaluare"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "Material": {
                "properties": {
                    "pdf": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Pdf"
                    },
                    "structurat": {
                        "anyOf": [
                            {
                                "items": {
                                    "$ref": "#/components/schemas/StructuraMaterial"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Structurat"
                    }
                },
                "type": "object",
                "title": "Material"
            },
            "Materiale": {
                "properties": {
                    "disciplina": {
                        "type": "string",
                        "title": "Disciplina"
                    },
                    "tip_material": {
                        "type": "string",
                        "title": "Tip Material"
                    },
                    "material": {
                        "$ref": "#/components/schemas/Material"
                    }
                },
                "type": "object",
                "required": [
                    "disciplina",
                    "tip_material",
                    "material"
                ],
                "title": "Materiale"
            },
            "Proba": {
                "properties": {
                    "tip_proba": {
                        "type": "string",
                        "title": "Tip Proba"
                    },
                    "pondere": {
                        "type": "number",
                        "title": "Pondere"
                    }
                },
                "type": "object",
                "required": [
                    "tip_proba",
                    "pondere"
                ],
                "title": "Proba"
            },
            "StructuraMaterial": {
                "properties": {
                    "titlu": {
                        "type": "string",
                        "title": "Titlu"
                    },
                    "descriere": {
                        "type": "string",
                        "title": "Descriere"
                    }
                },
                "type": "object",
                "required": [
                    "titlu",
                    "descriere"
                ],
                "title": "StructuraMaterial"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}